float to int64 in labels
CrossEntropyLoss expects labels to be of type torch.int64 (or torch.long)

+++ src/datasets/bert_dataset.py
@@ -42,6 +42,6 @@ class BERTDataset(Dataset):
             padding=True,
             return_tensors="pt",
         )
-        tokenized_batch["labels"] = torch.tensor(labels, dtype=torch.int)
+        tokenized_batch["labels"] = torch.tensor(labels, dtype=torch.int64)
 
         return tokenized_batch
